Gogo_Gemini/requirements.txt
google-generativeai
configparser
